+++ builders/ngsscbuild/models/options.ts
@@ -1,6 +1,7 @@
 import { JsonObject } from '@angular-devkit/core';
 
 export interface Options extends JsonObject {
+  additionalEnvironmentVariables: string[];
   aotSupport: boolean;
   browserTarget: string;
   ngsscEnvironmentFile: string;