+++ schematics/ng-add/schema.ts
@@ -1,4 +1,5 @@
 export interface Schema {
+  additionalEnvironmentVariables: string;
   /** Name of the project. */
   project: string;
   variant: 'process' | 'NG_ENV';